refactor: Cleanup GeoLayer.tsx
