+++ giraffe/src/utils/geo.ts
@@ -1,7 +1,9 @@
 import {GeoViewLayer} from '..'
+import {GeoQueryVariables} from '../types/geo'
 
 // Constants
 export const ZOOM_FRACTION = 8
+const circumEarth = 40075016.686
 
 export const getMinZoom = (width: number): number => {
   // Math.max(Math.log2(width/256)),Math.log2(height/256))
@@ -29,4 +31,18 @@ export const getRowLimit = (layers: GeoViewLayer[]) => {
       }
     })
   )
+}
+
+export const calculateVariableAssignment = (
+  width: number,
+  height: number,
+  lon: number,
+  lat: number,
+  zoom: number
+): GeoQueryVariables => {
+  const pixelRadius = Math.sqrt(width * width + height * height) / 2
+  const metersPerPixel =
+    (circumEarth * Math.abs(Math.cos((lat * Math.PI) / 180))) /
+    Math.pow(2, zoom + 8)
+  return {lon, lat, radius: (pixelRadius * metersPerPixel) / 1000}
 }
